Move CreateProject into tasks.go
